refactor: extract admin skill path projection
